user acc img save

+++ komax_app/models.py
@@ -25,33 +25,36 @@ class Worker(models.Model):
     )
 
     def save(self, *args, **kwargs):
-        image = Worker.objects.get(id=self.id).image
+        workers = Worker.objects.filter(id=self.id)
         old_path = None
-        print(bool(image))
-        if bool(image):
-            old_path = image.url
-            print(os.path.exists(old_path))
-            if os.path.exists(old_path):
-                os.remove(image.url)
+        if len(workers):
+            image = workers.first().image
+            # image = Worker.objects.get(id=self.id).image
 
-        image = Image.open(self.image)
-        image_format = self.image.name.split('.')[-1]
-        image_name = self.image.name.split('.')[0]
+            if bool(image):
+                old_path = image.url
+                print(os.path.exists(old_path))
+                if os.path.exists(old_path):
+                    os.remove(image.url)
 
-        output = BytesIO()
+        if bool(self.image):
+            image = Image.open(self.image)
+            image_format = self.image.name.split('.')[-1]
+            image_name = self.image.name.split('.')[0]
 
-        image = image.resize((200, 200))
-        if image_format == 'png':
-            image.save(output, format='PNG', quality=100)
-        else:
-            image.save(output, format='JPEG', quality=100)
-        output.seek(0)
+            output = BytesIO()
 
-        self.image = InMemoryUploadedFile(output, 'ImageField', '{name}.{format}'.format(name=image_name, format=image_format), 'image/jpeg', sys.getsizeof(output), None)
-        print(os.path.exists(old_path), old_path)
-        if old_path is not None and os.path.exists(old_path):
-            print("REMOVING OLD", old_path)
-            os.remove(old_path)
+            image = image.resize((200, 200))
+            if image_format == 'png':
+                image.save(output, format='PNG', quality=100)
+            else:
+                image.save(output, format='JPEG', quality=100)
+            output.seek(0)
+
+            self.image = InMemoryUploadedFile(output, 'ImageField', '{name}.{format}'.format(name=image_name, format=image_format), 'image/jpeg', sys.getsizeof(output), None)
+
+            if old_path is not None and os.path.exists(old_path):
+                os.remove(old_path)
         super(Worker, self).save(*args, **kwargs)
